Change to "set" philosophy
