+++ packages/bundle/src/boot/actual/component/fullDiff.ts
@@ -1,0 +1,14 @@
+// Additional
+export { default as AdaptiveCardContent } from '../../../adaptiveCards/Attachment/AdaptiveCardContent';
+export { default as AnimationCardContent } from '../../../adaptiveCards/Attachment/AnimationCardContent';
+export { default as AudioCardContent } from '../../../adaptiveCards/Attachment/AudioCardContent';
+export { default as HeroCardContent } from '../../../adaptiveCards/Attachment/HeroCardContent';
+export { default as OAuthCardContent } from '../../../adaptiveCards/Attachment/OAuthCardContent';
+export { default as ReceiptCardContent } from '../../../adaptiveCards/Attachment/ReceiptCardContent';
+export { default as SignInCardContent } from '../../../adaptiveCards/Attachment/SignInCardContent';
+export { default as ThumbnailCardContent } from '../../../adaptiveCards/Attachment/ThumbnailCardContent';
+export { default as VideoCardContent } from '../../../adaptiveCards/Attachment/VideoCardContent';
+export { default as ReactWebChat } from '../../../FullReactWebChat';
+
+// Overriding
+export { default as Composer } from '../../../FullComposer';
