+++ packages/bundle/src/boot/actual/hook/fullSet.ts
@@ -1,7 +1,6 @@
 export * from './minimalSet';
 
-// Additional
-export { default as useAdaptiveCardsHostConfig } from '../../../adaptiveCards/hooks/useAdaptiveCardsHostConfig';
-export { default as useAdaptiveCardsPackage } from '../../../adaptiveCards/hooks/useAdaptiveCardsPackage';
-export { default as useStyleOptions } from '../../../hooks/useStyleOptions';
-export { default as useStyleSet } from '../../../hooks/useStyleSet';
+export * from './fullDiff';
+
+// Explicit overrides.
+export { useStyleOptions, useStyleSet } from './fullDiff';
