+++ packages/bundle/src/boot/exports/component.ts
@@ -1,0 +1,1 @@
+export * from '../actual/component/fullSet';
