+++ packages/bundle/src/boot/exports/hook.ts
@@ -1,0 +1,1 @@
+export * from '../actual/hook/fullSet';
