+++ packages/bundle/src/boot/exports/index.old.ts
@@ -1,0 +1,22 @@
+// export * from '../actual/fullDiff';
+// export * from './minimal';
+
+// // Explicit overrides.
+// export {
+//   createStyleSet,
+//   default,
+//   ReactWebChat,
+//   renderWebChat,
+//   StrictStyleOptions,
+//   StyleOptions,
+//   version
+// } from '../actual/fullDiff';
+
+// // Backward compatibility.
+// // Web devs should move to named exports instead.
+// export * as decorator from './decorator';
+// export * as internal from './internal';
+
+// // Explicit overrides with backward compatibilty.
+// export * as Components from './component';
+// export * as hooks from './hook';
