+++ packages/bundle/src/boot/exports/minimal.old.ts
@@ -1,0 +1,12 @@
+// // Default is not re-exported when using `export *`.
+// import ReactWebChat from '../actual/minimalSet';
+
+// export * from '../actual/minimalSet';
+// export default ReactWebChat;
+
+// // Backward compatibility.
+// // Web devs should move to named exports instead.
+// export * as Components from '../actual/component/minimalSet';
+// export * as decorator from './decorator';
+// export * as hooks from '../actual/hook/minimalSet';
+// export * as internal from './internal';
