+++ packages/bundle/src/boot/exports/minimal.ts
@@ -1,4 +1,4 @@
-export * from '../actual/fullSet';
+export * from '../actual/minimalSet';
 
 // Backward compatibility.
 // Web devs should move to named exports instead.
@@ -6,5 +6,5 @@ export * as decorator from '../actual/decorator';
 export * as internal from '../actual/internal';
 
 // Explicit overrides with backward compatibilty.
-export * as Components from '../actual/component/fullSet';
-export * as hooks from '../actual/hook/fullSet';
+export * as Components from '../actual/component/minimalSet';
+export * as hooks from '../actual/hook/minimalSet';
