+++ packages/bundle/src/boot/iife/webchat.old.ts
@@ -1,0 +1,13 @@
+// import './polyfill/modern';
+
+// import * as actual from '../exports/index';
+
+// // "./full" is IIFE, just importing it will inject to globalThis.
+// import './webchat-minimal';
+
+// // Built on top of ./minimal.
+// // ./minimal export /decorator, /internal, /middleware, so we don't need to duplicate that behavior here.
+// window['WebChat'] = Object.freeze({
+//   ...window['WebChat'],
+//   ...actual
+// });
