+++ packages/bundle/tsup.config.ts
@@ -19,6 +19,8 @@ const commonConfig = applyConfig(config => ({
   ...config,
   entry: {
     'botframework-webchat': './src/boot/exports/index.ts',
+    'botframework-webchat.component': './src/boot/exports/component.ts',
+    'botframework-webchat.hook': './src/boot/exports/hook.ts',
     'botframework-webchat.internal': './src/boot/exports/internal.ts',
     'botframework-webchat.middleware': './src/boot/exports/middleware.ts'
   },
